pyproject.toml
[project]
name = "flight-booking-hw3"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrew",email = "aikhramov@edu.hse.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "grpcio (>=1.78.0,<2.0.0)",
    "grpcio-tools (>=1.78.0,<2.0.0)",
    "protobuf (>=6.31.1,<7.0.0)",
    "redis (>=5.0.4,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
